integrate mnest.py into nestle.py
'multi' method not yet working, but 'single' mode (default) should be
identical to previous behavior, with one caveat: the ellipsoidal volume
enlargement was previously hard-coded to 1.06^n, where n is the number
of dimensions. It is now an undocumented kwarg with a default value of 1.5.
set `enlarge=1.06**n` to achieve previous behavior.