parallelization
Adds functionality (based on changes from the `parallel` branch of `nestle`)
that enables points to be proposed in parallel using a provided Pool. Designed
to utilize `ThreadPoolExecutor`. Based on code from @aarchiba.

This commit also removes the `'classic'` MCMC option from the sampler for now.
This will be added back in later.

Finally, this also implements a few small changes to how weights are reported
to keep them in log space. I'll be coming back to this in a bit when I start
revamping things.