cleanup
This commit is designed to clean up the original `nestle` code both
in terms of formatting, style, etc. as well as in terms of implementation.
These changes are designed to lay the foundation for many of the upcoming
commits. A summary of (hopefully most) changes is below.

- The code is now PEP8 compatible.
- Formatting in descriptions has been homogenized. The descriptions
themselves have been changed where I felt some more/less detail was
needed.
- Added a few utility functions to do common checks (is the point in the
unit cube?, is the point in the ellipsoid?, etc.)
- Modified variable names in the spirit of making them more
self-descriptive wherever possible or to be more in line with nested
sampling terminology. Being precise will be important once we start moving
to dynamically adding single-point runs.
- Added the ability to supply additional variables involved in the
multi-ellipsoidal decomposition algorithm, in case users want to tweak
that by hand.
- `sample_ellipsoids` now returns a point and the number of ellipsoids `q` it
is contained in. This allows us to put the `p=1./q` rejection step in
elsewhere in the code, which will be useful when/if we add
importance sampling.
- Added an `nqueue` variable in `Sampler` to keep track of how the queue's
doing in parallel applications, which might be useful in future tests.
- Fixed a bug where `it` was actually 1 iteration behind when computing
`pointvol` (making the minimum volume slightly too large).
- Changed ordering of updates in the main nested sampling loop to emphasize
the difference between "live" points (which sample) and "dead" points
(which we sum over).
- Added the unit cube samples to the output `Results` under `'samples_unit'`.
This will be useful in visualizing runs once the proposal states can be
stored.
- Added `nlive` (the new `npoints`) to `Results` and summary outputs.