refactoring
This resolves #5 by refactoring the original `dynesty` code into a series
of individual pieces to make development and usage more modular. A summary
of all the changes are below:

- `__init__.py` imports `dynesty` directly and imports `ellipsoid` and `utils`
one step removed.
- `dynesty.py` now contains the main nested sampling loop along with the
`Results` class.
- `ellipsoid.py` contains all things ellipsoidal. This includes the
`Ellipsoid` and new `MultiEllipsoid` classes, along with helpful
functions for computing bounding ellipsoids and decompositions.
- `fakepool.py` contains the `FakePool` and `FakeFuture` classes used for
serial evaluation.
- `utils.py` contains a variety of utilities that used to awkwardly live
at the beginning of the huge chunk of code.
- `sampler.py` contains the base `Sampler` class. Things have been slightly
modified so that now the entire proposal process is sent to each member of
the provided `Pool`. Although this likely increases overhead by having to
transfer the sampler object, it allows local iterative proposals that
evolve deterministically and involve multiple likelihood calls.
- `globalsampler.py` now contains all samplers that work by bounding the
prior volume. This currently just is the `SingleEllipsoidSampler` and
`MultiEllipsoidSampler`.
- `localsampler.py` theoretically will contain all samplers that work by
trying to "move away" from live points.