quadratic integration
Implemented a quadratic integration scheme over the original linear one. This
should reduce our integration error to `O(N^-2)` at the cost of some
additional overhead. This resolves #1.

As part of this change, I also went through and fixed a few small
bugs in the original integration scheme, such as the use of `vol`
rather than `dvol` when computing the weights (they're close but not
the same thing) and a slightly incorrect summation of the final set of live
points (again, implementation was close but not exactly correct).