func to sampler
This commit resolves #12. The main nested sampling loop now is a function
that can be called from the base `Sampler` class similar to how
`emcee` does this. This enables users to keep track of the current state,
to start/stop runs on demand, and to continue previous runs.
The stopping criteria has consequently also been moved from the main interface
to this function to allow users to add more nested samples with different
stopping criteria if they like (e.g., running multiple runs with fixed
numbers of iterations). Results are stored in the same format as previous and
can be accessed via the `get_results()` method.

To accompany this change, I've split up the previous `sample` function into an
equivalent function that runs within the `Sampler` base class as well as a
`NestedSampler` interface that emulates the previous `nestle` interface but
instead returns a `Sampler` subclass instead of a `Result` class. I've also
done some work modifying the `Sampler` base class and subclasses to support
all these changes and (hopefully) mesh better together.

There currently aren't a lot of hooks in the base `Sampler` class to deal
with these changes; I hope to progressively add these in future commits as
I resolve more open issues.