results change
I've added a bunch of new quantities to the results output from a run. The
`Results` class has also been moved to it's own place now `results.py` to
facilitate possible substantial changes in the future (such as adding simple
plotting utilities, etc.).

I've also implemented a few small changes to the baseline sampler under the
hood. The `Sampler` can now be reset with the `reset` function, which avoids
having to instantiate another class. In addition, I've moved the final part of
the nested sampling loop (where we add the remaining live points) to
be a separate internal function call `_add_live_points`. While doing so,
I also disovered a slight bug in the implementation that meant we weren't
adding live points in the proper order. This should now be fixed.