refactor samplers
Resolves #15 by allowing sampling methods and bounding methods to be
specified separately. Currently only `uniform` choices are allowed for
sampling and `single` and `multi` ellipsoids for bounding.

The API for `dynesty.py` has been slightly modified to account for the new
changes, which now can take arguments to specify the bounding method and
the sampling method (conditioned on the bounds). The `Sampler` subclasses
now have separate functions that implement each of the sampling methods. These
are mostly placeholders that will be modified in the near-future. Finally,
the base `Sampler` class has been modified to allow for likelihood constraints
to be included when proposing a new point, which will be used by most other
samplers.