multi rwalk bugfix
Fixed a bug in the multi-ellipsoidal implementation of `'randomwalk'` where
I accidentally broke detailed balance.

There were two problems. The first is sometimes we ended up
outside of all the ellipsoids. Since we need to pick an ellipsoid to condition
on (rather than just having it specified as in the `'none'` and `'single'`
cases), we actually want to keep track of these things in `'multi'`.
Originally, I just ignored this and stuck with the original proposal. Now I
actually trigger a full update of the ellipsoids. This indicates we're missing
prior volume in our decomposition, which is Very Bad. Gives me more incentive
to implement the bootstrap volume fix so hopefully this rarely happens.

The second problem (discovered talking with @bd-j) was that I was
switching proposals after each iteration and (1) not including that proposal
asymmetry in the acceptance criterion and (2) gave a non-zero possibility that
we couldn't return to our starting position at all. There's a proper way to
do this, which I did not implement here, and a simple way, which I chose for
now. I'll document them both for posterity here.

The "proper" way would be to actually define a position-dependent proposal
distribution `q(x*|x)` based on a mixture of all ellipsoids located at `x`.
We can propose `x*` in two steps: first, sample an ellipsoid `q(i|x)` that
overlaps with `x` uniformly, and then compute `x*=x+dx` by sampling `dx` from
the stationary ellipsoid proposal distribution (i.e. centered on `x`). The
proposal `q(x*|x)` can then be computed by transforming the relevant `dx(i)`
vectors associated with each ellipsoid back to the unit n-sphere,
computing `q(x*|x,i)` (this is zero if outside the ellipsoid/sphere),
and simply adding up the `q(x*|x,i)`'s over the ellipsoids. The latter
procedure can also be used to compute the reverse proposal distribution
`q(x|x*)` at `x*`. We then use this to modify the acceptance criterion
accordingly. This proposal would allow for points to random walk within
contiguous structures bounded by the ellipsoids (and even between very nearby
clusters).

A simpler option is to say that we condition our random walk on one ellipsoid
at a time. This means that we effectively try and sample uniformly within
the likelihood contour bounded by that ellipsoid and that ellipsoid only.
This is less efficient in the sense that we are not sampling uniformly across
the contiguous region bounded by the ellipsoids, but since we (1) select
a point uniformly and then (2) select an ellipsoid uniformly, I believe
we still are generating independent, uniform samples from the prior.

On a sidenote, I also changed the name of `samplers.py` to `nestedsamplers.py`
since it was starting to confuse collaborators.