bootstrap enlargement factor
Resolves #9! The enlargement factor for ellipsoid volumes is now calculated
during the course of an update by bootstrap resampling the set of live points,
computing the distances away from the set of "missing" points (relative to
the closest ellipsoid), and then taking the maximum from the whole set
of distances. This process can be controlled via the `bootstrap` keyword. The
`enlarge` option now is by default set to 1 unless `bootstrap = 0`, at which
point we set `enlarge=1.25` (a little more conservative than the original
option in `nestle`).

Some additional notes:
- this commit reverts a change where I attempted to be clever with
`linalg.norm` but didn't choose the right norm, lead to a horrendous
underestimate of the ellipsoid volumes (and subsequently gargantuan
ellipsoids).
- since sometimes `kmeans2` throws a warning when one of the clusters is
unnecessary (which can happen during resampling), I now just suppress all
warnings when running it.
- added an option to dump proposals when doing `run_nested` since it's not
crucial that those are saved.