balls and cubes
Resolves #8, where I've now implemented the RADFRIENDS and SUPFRIENDS bounding
methods from [Buchner (2014)](https://arxiv.org/abs/1407.5459). They can
be called under the `balls` and `cubes` bounding keywords.

In more detail, these simply constrain the proposals to be a set of `K`
n-balls or n-cubes centered on each of the live points, where live points
are selected by the volume of their union. Sampling from them is thus
equivalent to sampling from a set of possibly overlapping ellipsoids in
`'multi'`, in the sense that we only accept points with probability `p=1/q`
where `q` is the number of cubes/balls that a proposed point `u` is contained
in. This entails a check over `K` live points instead of `M` ellipsoids,
although the distance computations are a bit quicker (I think?).

The vanilla implementation is not very efficient for two reasons. First, by
construction there end up being a lot of overlapping balls/cubes, meaning
that tons of proposed points don't get accepted. Second, my current
implementation for the bootstrap step is very inefficient, and involves
checking `O(K^2)` (in the leave-one-out case) distances and `O(NKJ)` in the
bootstrap update, where `J` is expected number of points left out at each
iteration. Both of these scale quite strongly with `K`. I think the correct
thing to do would be to do this in a tree-based way, but given that the
`multi` method is now more robust anyways thanks to the bootstrap update
improving this isn't a high priority.