bounding changes
Refactored the `NestedSampler` objects so that they all utilize
`bounding` objects from the new `bounding.py` module (renamed from
`ellipsoid.py`). This allows me to call the bounding methods separately
on a "as needed" basis, which will be useful for dynamic sampling.

As an added bonus, this also simplifies the layout of each `NestedSampler`. It
also gives a common structure that I should be able to take advantage when I
start implementing standard plotting functions.