initial DynamicSampler commit
This commit pushes forward the first iteration of the `DynamicSampler`, which
uses the dynamic sampling scheme I first referenced in #23 to sample
adaptiely based on some user-specified `wt_function` (the default just
is a weighted average between the evidence or posterior weight).

The `DyanamicSampler` still has a bunch of issues that I need to work out
with regards to properly estimating prior volumes, errors, etc. These should
be worked out in the coming days now that I've finished deriving a bunch of
results offline. To summarize the basics though, the `DynamicSampler` is a
hack over the previous `NestedSampler`. Using a baseline run, it derives an
initial approximation for the bounding proposals and the underlying posterior
mass. It then iteratively proposes new points within likelihood bounds
derived from previous samples using a provided weight function and
iterates until some convergence criteria are met.

There are a bunch of other small changes in this commit based on feedback from
several users (with special thanks to @bacook17). A quick rundown:
- The remaining `dlogz` now is computed by default and output in real time
so users can check convergence. The current efficiency is also returned and
displayed.
- The `sys.stderr` print messages now properly update.
- Several bugs in the plotting utilities have now been fixed.
- Bounding objects now by default compute their "union" volume by Monte Carlo
integration (e.g., `mell.vol`) as well as overlap with the unit cube (e.g.,
`mell.funit`). These are helpful for debugging and already have shown that
the `'multi'` bootstrap can be unstable for small `K` live points (which might
require adjusting splitting defaults to be even more conservative).
- Samplers now keep track of the proposal used at a given iteration directly.
This was a useful change for keeping track of things in the the
`DynamicSampler` which can jump around among proposal indices.