error fixes
Fixes a bunch of issues involved with nested sampling errors. Resolves #28.

Issues with noise in the `h`-based (KL divergence/"information") `logz` error
estimates have been mitigated by moving to a more natural variance default
and using the natural expected change in log(volume) (which is what the error
estimate is based on).

The simulated and resampling-based errors (and the combined errors) have
now been updated to deal with dynamic runs (and also to integrate with
trapezoids). For simulating errors, this involves some additional bookkeeping
and a set of loops over each set of uniform order statistics we're probing
(to include covariances). Results look about as expected: posteriors get
significantly more noisy once you account for statistical/sampling
uncertainties, while the evidence is about as stable as can be expected.