resample_run bugfix
Fixes an issue associated with resampling a run connected to issue #43.

The actual problem was subtle and didn't show up until further testing after
combining `resample_run` and `simulate_run`, which showed that the prior
volumes associated with `resample_run` were actually incorrect. After some
digging, it turns out that this is due to the specific way we need to combine
repeated strands: a strand that is resampled `N` times actually should be
treated as a single nested sampling run with a constant number of live points
`N`. This makes no difference if the number of live points is constant
throughout the run, but changes results if we added the final sample
from each strand (i.e. we "recycled" the final live point). In that case,
the final set of `N` points should be treated as having a decreasing number
of live points (`N`, `N-1`, `N-2`, ...`). I had previously been ignoring this
subtle edge-case (I had assumed they stayed as `N` live points throughout).

This commit tries to deal with the above edge case in an function-agnostic way.
A flag has also been added in the simple case where the user has utilized a
constant number of live points without any final "recycling".