bounding bugfixes
More small bugs that have now been fixed. Summary of changes is below.
- Ellipsoids now automatically raise a `ValueError` if the eigenvalue/vector
decomposition fails. I haven't yet identified all the exact times when this
actually occurs, but this should at least allow for debugging (especially when
using a `pool`).
- The `within` method for `MultiEllipsoid`, `RadFriends` and `SupFriends` has
now been slightly altered to avoid issues with (some) older versions of `numpy`.
- If sampling either fails or stalls during when sampling using slices or
random walks, a `RuntimeError` is raised along with a host of internal values.
This should allow users (but mostly me) to do some debugging.
- When proposing an update within the `MultiEllipsoidSampler` in the case where
the point lies outside of all the bounding ellipsoids, an automatic update
is triggered. The default log-volume associated with the `pointvol` passed to
this update has been changed to more accurately reflect the current state of
the sampler.