kde plotting
Added in the ability to plot the importance weight **PDF** as a function of
log-volume, as opposed to just the importance weights themselves. Resolves #52
and the issues discussed there. In addition, I've also by default normalized
the log-likelihoods when plotting to avoid numerical underflow/overflow
errors. I don't think any users will be particularly peeved since it's the
general trend that's most important (and if they want the actual value
they can just grab them from the `results`).