small bugfixes
- Resolves #72 by changing `results` to `res`. I haven't changed the
underlying tuple structure, but I'll think about whether it's worth
changing around the internals so that users only have to access the
`sampler.results` dictionary instead of memorizing/copying a bunch of
variable names.

- Resolves #75 by multiplying the normalizing factor used to rescale
`scale` in `update_rwalk` by the dimension of the prior. This means
that we now rescale in volume rather than length-scale, which should
cut down on some of the variation in `rwalk` during runtime, although it
might make the initial adjustments a little slower.

- Added some stuff to the FAQ.