batch generator update
This update should now resolve #77 properly by yielding the newly
initialized live points within the context of an external generator. To
properly ignore these "invisible" points (which are dealt with during
sampling but by themselves don't mean anything upon initialization),
I have each point tracked by negative indices for `worst` which is passed
to the tuple.

**Note that this change may break backwards compatibility for previous users of
the generator framework with the dynamic nested sampler. Please
see the updated `run_nested` and `add_batch` code for an example of how
to deal with this using a very simple change.**