v0.8.4
This commit implements a number of small changes and bugfixes. Many thanks
to @guillochon and @catherinezucker for finding most of the bugs
and providing many of the suggestions that have gone into this update.

Summary:
- Version changed to v0.8.4.
- Parallelism has been updated so that users shouldn't need to specify
`queue_size` when passing a pool unless they want to manually override
the `pool.size` option. This resolves #79. Demos 1 and 2 have also been
updated accordingly (but not the docs).
- The algorithm used to ensure matrices are positive-definite (and thus
invertible) has been simplified. While it won't be as efficient, hopefully
this should resolve some of the bugs/edge cases that have come up. This
should resolve #78.
- I had to implement a small bugfix after merging with the recent pull
from @bacook17. This resolves issues where `add_batch` wasn't updating
properly in `run_nested`.
- `_quantile` has now been moved to `utils` and renamed `quantile` since
users will want to access it outside of plotting.
- Added a new `thin` argument to `cornerpoints` so that not all points
are always plotted and added some clipping to avoid generating
excessively intensive plots.