hslice
This commit adds in a hamiltonian slice sampling option (`'hslice'`) into
`dynesty`. This resolves (finally!) #13.
Internal testing shows that it appears to be more robust
than `'rslice'` for a fixed number of slices (although markedly less
efficient for obvious reasons). It has not been stress-tested on
pathological distributions but appears to perform alright on a 50-D Gaussian.
This should also resolve #68. I'll definitely return to this in the future
to see if some additional tuning can improve performance, but for now I'm
just glad that this seems to work.

Also, I changed the default number of slices to 5. 3 was a bit too aggressive.
I might increase this further for `'rslice'` and `'hslice'` in the future.