fast kld
This commit resolves #82 by introducing an `approx` argument
for `jitter_run` that allows the user to approximate all the
joint uniform order statistics as their Beta marginals. This
significantly speeds up the evaluation time, especially in
the context of `kld_error`.

In addition, I have changed the default stopping criteria
options in `dynesty` to take advantage of this change. The default
option is now `sim_approx`, which uses `jitter_run` (with
`approx=True`) but boosts the resulting stopping criteria by a
factor of two for the posterior and sqrt(2) for the evidence.
I've also increased `n_mc` accordingly to beat down
some of the shot noise in `stop_val`. Internal tests have shown this
performs comparably to the previous version that used `'simulate'`,
so hopefully this will be a good change going forward.