plot kde options
Added in kde options to both `traceplot` and `cornerpoints` mirroring
the implementation in `runplot`. This is just trying to ensure that plotting
results from the `dynamicsampler` look similar across all venues. The second
demo has also been updated accordingly.

I also just added a note to the README that points out that the documentation
is somewhat out of date on the newer options (the API is fine, but the actual
examples/details are a somewhat behind).