changed defaults
This commit resolves #87. The default options for a host of parameters has
been changed to enable more stable behavior for the average user. These are:
- `nlive` has been changed to `500` everywhere to enable better default
performance in moderate dimensions. This includes in `nlive_init` and
`nlive_batch` in `dynamicsampler`.
- `update_interval` has been increased to `0.8`.
- The `dlogz` formula for the static `sampler` has been modified to
`1e-3 * (nlive - 1) + 0.01` so that it is a better compromise between
evidence and posterior estimation. This gives `~0.5` when `nlive=500`,
similar to the previous `nestle` value, which I think is more appropriate.

I might revisit the sampling options in the near future to further accomodate
users who want to apply it right away to arbitrary-dimensional problems.