bugfixes
This commit finally gets around to (trying to) fix some issues
in the code. A quick summary is below:

- Ellipsoids were being over-regularized. Dependencies
on `sklearn` have been removed and a more conservative
approach has been implemented. Hopefully this should
fix this.

- `'rwalk'` was having issues when the bounding ellipsoid was
a significant fraction of the unit cube volume due to some poor
interactions. This should hopefully be fixed; `scale` now has an
reasonable upper bound imposed and is also adjusted if RNG is
really inefficient. Hopefully this should be enough...

- Small bugfixes in the dynamic sampler.