v0.9.3
Touches a lot of files, but actually not a lot of big changes. This mostly
locks in lots of the results from version 0.9.2b by updating
the documentation to be consistent with them. I will make an active effort
to update the documentation in the future when I start adding in more
features. Which will be imminent because it's finally time to start wrapping
this project up and get it out!

Summary of changes additional changes on top of 0.9.2b:
- Updated docs.
- Changed plotting defaults to 95% CI (gotta start somewhere).
- Fixed some pep8 compatibility stuff.
- Pushed to PyPI.