v0.9.5
Initial commit for version 0.9.5.

This adds support for periodic boundary conditions, which are now included as
part of the Exponential Wave example. This finally resolves #94, which had
been sitting unfinished on a local version of the code for months. The
internals have been slightly refactored to accomodate a new `unitcheck`
function that gets passed around. I also removed `random_choice`, which
essentially was never used and otherwise totally deprecated.