Removes the automatic reflective boundaries
In testing using bilby, we found that reflective boundaries can result in odd
behaviour at the boundary. We believe this is because the detailed balance of
the MCMC steps are no longer satisfied (multiple ways exist to get to the same
point). This removes the automated reflective boundaries for
non-periodic parameters (but retains periodic boundaries).