MInor fix to the reflectice implementation
While the previous implementation worked. It was still possible to go so
far out of the unit cube that it didn't reflect properly (i.e. it was
mapped to a point on the other side, outside, of the unit cube). This
fixes that and ensures all points are mapped back in properly.